requirements.txt
asarPy==1.0.1
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
mysql-connector-python==9.5.0
PyJWT==2.10.1
pymongo==4.16.0
python-dotenv==1.2.1
sqlparse==0.5.5
